created a base class for all image tasks based on ocio
