+++ src/lib/ingestor/Task/Image/__init__.py
@@ -1,5 +1,6 @@
 from .UpdateImageMetadata import UpdateImageMetadata
 from .ConvertImage import ConvertImage
+from .Ocio import Ocio, OcioConfigurationError
 from .ResizeImage import ResizeImage
 from .ImageThumbnail import ImageThumbnail
 from .ColorTransformation import ColorTransformation
